Move code to src/
