+++ src/account-db.js
@@ -17,7 +17,10 @@ function getAccountDb() {
     accountDb = openDatabase(dbPath);
 
     if (needsInit) {
-      let initSql = fs.readFileSync(join(__dirname, 'sql/account.sql'), 'utf8');
+      let initSql = fs.readFileSync(
+        join(__dirname, '../sql/account.sql'),
+        'utf8'
+      );
       accountDb.exec(initSql);
     }
   }
